chore(sanity): rename versions to bundles in core directory
