+++ packages/sanity/src/core/bundles/util/dummyGetters.ts
