+++ packages/sanity/src/core/studio/components/navbar/StudioNavbar.tsx
@@ -16,10 +16,10 @@ import {type RouterState, useRouter, useRouterState} from 'sanity/router'
 import {styled} from 'styled-components'
 
 import {Button, TooltipDelayGroupProvider} from '../../../../ui-components'
+import {GlobalBundleMenu} from '../../../bundles/components/GlobalBundleMenu'
 import {type NavbarProps} from '../../../config/studio/types'
 import {isDev} from '../../../environment'
 import {useTranslation} from '../../../i18n'
-import {GlobalBundleMenu} from '../../../versions/components/GlobalBundleMenu'
 import {useToolMenuComponent} from '../../studio-components-hooks'
 import {useWorkspace} from '../../workspace'
 import {ConfigIssuesButton} from './configIssues/ConfigIssuesButton'
